SYAL-760 refactor and javadocs
