+++ src/main/java/com/avispl/symphony/dal/avdevices/encoderdecoder/haivision/x4encoder/common/CreateOutputStreamMetric.java
@@ -1,7 +1,7 @@
 /*
  *  * Copyright (c) 2022 AVI-SPL, Inc. All Rights Reserved.
  */
-package com.avispl.symphony.dal.avdevices.encoderdecoder.haivision.x4encoder.dropdownlist;
+package com.avispl.symphony.dal.avdevices.encoderdecoder.haivision.x4encoder.common;
 
 /**
  * CreateOutputStream
