+++ src/main/java/com/avispl/symphony/dal/avdevices/encoderdecoder/haivision/x4encoder/common/OutputMonitoringMetric.java
@@ -37,6 +37,7 @@ public enum OutputMonitoringMetric {
 	LATENCY("Latency",false,false),
 	OCCURRED("Occurred",false,false	);
 
+
 	private final String name;
 	private boolean isNormalize;
 	private boolean isReplaceComma;
